backend/pyproject.toml
[project]
name = "darwin"
version = "0.1.0"
description = "Self-improving chess engine via agentic tournament selection"
requires-python = ">=3.11"
dependencies = [
  "chess>=1.11.0",
  "anthropic>=0.40.0",
  "google-genai>=1.0.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "sqlmodel>=0.0.16",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "websockets>=12.0",
  "modal>=1.4.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["darwin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
